# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rizki Mandiri API Documentation",
        "description": "L5 Swagger OpenApi description for Rizki Mandiri POS API",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Base API Server"
        }
    ],
    "paths": {
...
        "/api/inventory/{id}": {
            "put": {
                "tags": [
                    "Inventory Base"
                ],
                "summary": "Update an existing inventory record",
                "description": "Update the details of an existing inventory ledger entry",
                "operationId": "updateInventoryLedger",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the inventory ledger to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "type"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-25"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "sale",
                                        "enum": [
                                            "initial",
                                            "purchase",
                                            "sale",
                                            "sale_item"
                                        ]
                                    },
                                    "branch_id": {
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "item_name": {
                                        "type": "string",
                                        "example": "Item A",
                                        "nullable": true
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 5,
                                        "nullable": true
                                    },
                                    "unit_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20000,
                                        "nullable": true
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100000,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Entry updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Entry updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized action"
                    },
                    "404": {
                        "description": "Data not found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Inventory Base"
                ],
                "summary": "Delete an existing inventory record",
                "description": "Remove the specified inventory ledger entry from database",
                "operationId": "destroyInventoryLedger",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the inventory ledger to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entry deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Entry deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized action"
                    },
                    "404": {
                        "description": "Data not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}